Archive parity c driver and tests, #632 and #621 (#908)
* [C] Basic test for data packet dispatcher.

* [C] Test for correct handling of message without subscription.

* [C] Add stub transport bindings.  Add test that setup messages will be elicited.

* [C] Pull common functions into the test fixture.

* [C] Use conductor queue to verify conductor commands.  Add test for receiving setup message.

* [C] Finish unit test parity for data packet dispatcher.

* [C] Add unit tests for session specific subscriptions to data packet dispatcher.

* [C] Support remove_if on int64_to_ptr hash map.

* [C] Small touch ups for remove_if functionality.

* [C] Add aeron_int64_to_tagged_ptr_hash_map.

* [C] Get by session subscriptions working for the data packet dispatcher.

* [C] Implement int2int counter map.

* [C] Use new int2int counter map for handling endpoint reference counting.

* [C] Convert from int32 counter map to int64.

* [C] Use out parameter to receive old values from the map on update and allow potential errors to propagate.

* [C] Treat set value to initial_value as remove on put.

* [C] Start to integrate counter map for session specific ref counting into receive channel endpoint.

* [C] Remove registration_id from test code (always set to -1).

* [C] Correctly register and de-register session specific subscriptions.

* [C] Refactoring - extract unlink_from_endpoint function.

* [C] Ensure that links are matched correctly when session id is specified.

* [C] Move counter map into same location as other c map implementations.

* [C] Add shared aeron_map.{c,h}.  Centralise default load factor constant instead of defining it per map.

* [C] Use common aeron_map_compound_key.

* [C] Remove some noise.

* [C] Better error checking and refactoring.

* [C] Add aeron_receive_destination.  Introduce destination array into aeron_receive_channel_endpoint.  Move transport, channel, data_paths and current control address down to destination.

* [C] TODOs.

* [C] Add basic add_rcv_destination support to aeron_driver_conductor.  Pass messages to receive proxy (no implementation in receiver).

* [C] Add basic remove_rcv_destination support to aeron_driver_conductor.  Pass messages to receive proxy (no implementation in receiver).

* [C] Add aeron_udp_channel_equals function for use in destination removal.

* [C] Fix link error.

* [C] Support add/remove for destinations in endpoints and publication images.

* [C] Send status messages to the list of destinations held in the publication image.

* [C] Handle eos on heartbeats across all destination for an image.  Fix up logic for determining if a destination has a pre-specified control address.  Use the correct control address when sending NAKs.

* [C] Correctly track active transport count using time of last frame.

* [C] Allow tracking of connections on narrow under runs.

* [C] Add destination to receiver (poller, endpoint, images).

* [C] Complete add/remove destination.  Create appropriate linkage between endpoint and destination transport.

* [C] Handle adding destination on setup from data packet dispatcher.

* [C] Increment the counters correctly.

* [C] Fix up re-resolution and error reporting from conductor.

* [C] Implement tag handling of publication session ids.

* [C] Return more useful error message if a client tries to re-use a channel endpoint while it is closing.

* [C] Fix multiple memory leaks trigger from within driver conductor tests.

* [C] Fix multiple memory leaks triggered from udp_channel_test.

* [C] Free aeron_uri from destination tracker when destination removed.

* [C] Fix memory leaks in data packet dispatcher test.

* [C] Fix memory leaks in publication image test.

* [C] Handle freeing URI by passing it back to the conductor thread.

* [C] Remove memory leaks from transport loss test.

* [C] Remove memory leaks from flow control test.

* [C] Merge fixes and spelling.

* [C] Simplify address/port formatting.

* [C] Formatting.

* [C] Remove TODO comment.

* [C] Ensure the correct term length value is used when using a tag for the session-id in the URI on addPublication.

* [C] Make sure when a publications sees a connection that it already has interest in, but doesn't have a connection for that destination, add it.

* [Java] Working system tests for MDS on the C driver.

* [Java] Enable re-resolution for MDS with the C driver.

* [Java] In BasicArchiveTest switch to using separate Archive and Media Driver processes.

* [Java] ArchiveTest and BasicArchiveTest passing with C Media Driver.

* [Java] ReplicateRecordingTest passing with C media driver.

* [Java] ReplayTest passing with C media driver.

* [Java] ArchiveAuthenticationTest passing with C media driver.

* [Java] ExtendRecordingTest passing with C media driver.

* [Java] ManageRecordingHistoryTest passing with C media driver.

* [Java] Handle resource temporarily unavailable as a warning on registration.  Remove more cases of the C driver not supported.  Add a Tests.reAddSubscription method to handle exception retry.

* [Java] More useful failure message.

* [Java] Remove experimental test.

* [Java] Remove extended timeouts used for debug.

* [Java] Checkstyle.

* [C] Add missing parameter for non-linux networking code.

* [C] Move header into extern "C" block to prevent compilation error.

* [C] Add <array> include to fix compile error on MacOS and Windows.

* [C] Fix test error on Windows.  Load function pointer for comparison from dll.

* [C] Fix warnings for Windows build.

* [C] Fix compile error for non-Windows platforms.

+++ aeron-system-tests/src/test/java/io/aeron/archive/ArchiveTest.java
@@ -23,10 +23,10 @@ import io.aeron.archive.client.RecordingEventsAdapter;
 import io.aeron.archive.codecs.SourceLocation;
 import io.aeron.driver.MediaDriver;
 import io.aeron.driver.ThreadingMode;
-import io.aeron.driver.status.SystemCounterDescriptor;
 import io.aeron.logbuffer.FragmentHandler;
 import io.aeron.logbuffer.FrameDescriptor;
 import io.aeron.logbuffer.Header;
+import io.aeron.test.TestMediaDriver;
 import io.aeron.test.Tests;
 import org.agrona.*;
 import org.agrona.collections.MutableBoolean;
@@ -85,7 +85,7 @@ public class ArchiveTest
     private String publishUri;
     private Aeron client;
     private Archive archive;
-    private MediaDriver driver;
+    private TestMediaDriver driver;
     private long recordingId;
     private long remaining;
     private int messageCount;
@@ -107,6 +107,11 @@ public class ArchiveTest
 
     private void before(final ThreadingMode threadingMode, final ArchiveThreadingMode archiveThreadingMode)
     {
+        if (threadingMode == ThreadingMode.INVOKER)
+        {
+            TestMediaDriver.notSupportedOnCMediaDriverYet("Foo");
+        }
+
         rnd.setSeed(seed);
         requestedInitialTermId = rnd.nextInt(1234);
 
@@ -129,7 +134,7 @@ public class ArchiveTest
         requestedStartPosition =
             ((requestedStartTermId - requestedInitialTermId) * (long)termLength) + requestedStartTermOffset;
 
-        driver = MediaDriver.launch(
+        driver = TestMediaDriver.launch(
             new MediaDriver.Context()
                 .termBufferSparseFile(true)
                 .threadingMode(threadingMode)
@@ -138,18 +143,22 @@ public class ArchiveTest
                 .errorHandler(Tests::onError)
                 .dirDeleteOnStart(true));
 
-        archive = Archive.launch(
-            new Archive.Context()
-                .maxCatalogEntries(MAX_CATALOG_ENTRIES)
-                .fileSyncLevel(SYNC_LEVEL)
-                .mediaDriverAgentInvoker(driver.sharedAgentInvoker())
-                .deleteArchiveOnStart(true)
-                .archiveDir(new File(SystemUtil.tmpDirName(), "archive-test"))
-                .segmentFileLength(segmentFileLength)
-                .threadingMode(archiveThreadingMode)
-                .idleStrategySupplier(YieldingIdleStrategy::new)
-                .errorCounter(driver.context().systemCounters().get(SystemCounterDescriptor.ERRORS))
-                .errorHandler(driver.context().errorHandler()));
+        final Archive.Context archiveContext = new Archive.Context()
+            .maxCatalogEntries(MAX_CATALOG_ENTRIES)
+            .fileSyncLevel(SYNC_LEVEL)
+            .deleteArchiveOnStart(true)
+            .archiveDir(new File(SystemUtil.tmpDirName(), "archive-test"))
+            .segmentFileLength(segmentFileLength)
+            .threadingMode(archiveThreadingMode)
+            .idleStrategySupplier(YieldingIdleStrategy::new)
+            .errorHandler(Tests::onError);
+
+        if (threadingMode == ThreadingMode.INVOKER)
+        {
+            archiveContext.mediaDriverAgentInvoker(driver.sharedAgentInvoker());
+        }
+
+        archive = Archive.launch(archiveContext);
 
         client = Aeron.connect();
 
@@ -171,8 +180,14 @@ public class ArchiveTest
 
         CloseHelper.closeAll(client, archive, driver);
 
-        archive.context().deleteDirectory();
-        driver.context().deleteDirectory();
+        if (null != archive)
+        {
+            archive.context().deleteDirectory();
+        }
+        if (null != driver)
+        {
+            driver.context().deleteDirectory();
+        }
     }
 
     @ParameterizedTest
